Seperate metadata code and add variable of each key
